requirements.txt
streamlit>=1.30
pytest>=7.0
google-genai
python-dotenv